refactor(claude): isolate alternate config service factory
