+++ config/tsconfig.cli.json
@@ -18,6 +18,7 @@
     "../src/main/antigravity/hook-service.ts",
     "../src/main/claude/claude-config-dir-discovery.ts",
     "../src/main/claude/claude-config-dir-hook-controls.ts",
+    "../src/main/claude/claude-config-dir-hook-service.ts",
     "../src/main/claude/hook-settings.ts",
     "../src/main/claude/hook-service.ts",
     "../src/main/claude/statusline-script.ts",
