+++ src/main/agent-hooks/remote-managed-hook-installers.ts
@@ -1,7 +1,8 @@
 import type { SFTPWrapper } from 'ssh2'
 import type { AgentHookInstallStatus, AgentHookTarget } from '../../shared/agent-hook-types'
 import { ampHookService } from '../amp/hook-service'
-import { claudeHookService, createClaudeConfigDirHookService } from '../claude/hook-service'
+import { createClaudeConfigDirHookService } from '../claude/claude-config-dir-hook-service'
+import { claudeHookService } from '../claude/hook-service'
 import { discoverRemoteClaudeConfigDirNames } from '../claude/claude-config-dir-discovery'
 import { codexHookService } from '../codex/hook-service'
 import { geminiHookService } from '../gemini/hook-service'
