+++ src/main/claude/hook-service.ts
@@ -40,7 +40,6 @@ import {
   removeManagedStatusLine,
   type ClaudeCompatibleHookSettings
 } from './hook-settings'
-import { isClaudeFlavorConfigDirName } from './claude-config-dir-discovery'
 
 type ClaudeHookServiceOptions = {
   agent: AgentHookInstallStatus['agent']
@@ -350,18 +349,3 @@ export class ClaudeHookService {
 }
 
 export const claudeHookService = new ClaudeHookService()
-
-/** Service for a discovered `~/.claude-<name>` / `~/.claude.<name>` config
- *  dir. Same agent id and shared claude-hook script — these dirs run genuine
- *  claude posting to /hook/claude; only the settings.json location differs. */
-export function createClaudeConfigDirHookService(configDirName: string): ClaudeHookService {
-  if (!isClaudeFlavorConfigDirName(configDirName)) {
-    // Why: this value becomes a home-relative write path, so validate it again at the factory boundary.
-    throw new Error('Invalid Claude config-dir name')
-  }
-  return new ClaudeHookService({
-    agent: 'claude',
-    displayName: 'Claude',
-    settings: { configDirName, scriptBaseName: 'claude-hook' }
-  })
-}
